requirements.txt
Flask==2.3.3
langchain==0.1.12
pypdf==3.17.1
fastembed==0.2.4
openai==1.3.6
langchainhub==0.1.14
qdrant-client==1.10.1
python-dotenv==1.0.0
